refactor(ports): reorganize and update audio and subtitle processor imports
