+++ app/adapters/open_subtitles.py
@@ -1,7 +1,7 @@
 from typing import List
 from app.integration.http_client import HttpClient
 from app.core.config import settings
-from app.core.ports import SubtitleMovies
+from app.core.common.ports import SubtitleMovies
 from app.core.movies.application.dto.movie_dto import MovieSearchOut, MovieOut
 from datetime import datetime
